localstack-typedb/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "plux>=1.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "localstack-typedb"
version = "0.1.0"
description = "LocalStack Extension: TypeDB on LocalStack"
readme = {file = "README.md", content-type = "text/markdown; charset=UTF-8"}
requires-python = ">=3.9"
authors = [
    { name = "LocalStack + TypeDB team"}
]
keywords = ["LocalStack", "TypeDB"]
classifiers = []
dependencies = [
    "httpx"
]

[project.urls]
Homepage = "https://github.com/whummer/localstack-utils"

[project.optional-dependencies]
dev = [
    "boto3",
    "jsonpatch",
    "localstack",
    "pytest",
    "rolo",
    "ruff",
    "typedb-driver",
]

[project.entry-points."localstack.extensions"]
localstack_typedb = "localstack_typedb.extension:TypeDbExtension"
